+++ lib/assets/javascripts/cartodb/new_dashboard/background_importer/background_importer_view.js
@@ -2,6 +2,7 @@ var $ = require('jquery');
 var cdb = require('cartodb.js');
 var ImportsCollection = require('../../new_dashboard/background_importer/imports_collection');
 var ImportItem = require('../../new_dashboard/background_importer/background_importer_item_view');
+var ImportsModel = require('../../new_common/imports_model');
 
 /**
  *  Background importer view
@@ -32,7 +33,10 @@ module.exports = cdb.core.View.extend({
     this.collection.bind('add', this._addImport, this);
     this.collection.bind('add remove', this._checkCollectionSize, this);
     this.collection.bind('change add remove reset', this._checkImports, this);
+    cdb.god.bind('datasetSelected remoteSelected', this._addDataset, this);
+    cdb.god.bind('fileDropped', this._onDroppedFile, this);
     this.add_related_model(this.collection);
+    this.add_related_model(cdb.god);
   },
 
   _addImport: function(m) {
@@ -111,8 +115,18 @@ module.exports = cdb.core.View.extend({
     this.collection.destroyCheck();
   },
 
-  add: function(imp) {
-    if (imp) { this.collection.add(imp) }
+  _addDataset: function(d) {    
+    if (d) {
+      var imp = new ImportsModel({ upload: d }, { user: this.user });
+      this.collection.add(imp);
+    }
+  },
+
+  _onDroppedFile: function(files) {
+    if (files) {
+      var imp = new ImportsModel({ upload: { type: 'file', value: files } }, { user: this.user })
+      this.collection.add(imp);
+    }
   },
 
   show: function() {
