+++ lib/assets/javascripts/cartodb/new_dashboard/background_importer/imports_collection.js
@@ -1,7 +1,7 @@
 var Backbone = require('backbone');
 var _ = require('underscore');
 var pollTimer = 30000;
-var ImportsModel = require('../../new_dashboard/background_importer/imports_model');
+var ImportsModel = require('../../new_common/imports_model');
 
 /** 
  *  Imports collection
