+++ lib/assets/test/spec/cartodb/new_dashboard/background_importer/background_importer_item_view.spec.js
@@ -1,6 +1,6 @@
 var cdb = require('cartodb.js');
 var Router = require('../../../../../javascripts/cartodb/new_dashboard/router');
-var ImportsModel = require('../../../../../javascripts/cartodb/new_dashboard/background_importer/imports_model');
+var ImportsModel = require('../../../../../javascripts/cartodb/new_common/imports_model');
 var BackgroundImporterItemView = require('../../../../../javascripts/cartodb/new_dashboard/background_importer/background_importer_item_view');
 
 describe('new_dashboard/background_importer/background_importer_item_view', function() {
