+++ lib/assets/test/spec/cartodb/new_dashboard/background_importer/background_importer_view.spec.js
@@ -1,7 +1,7 @@
 var Backbone = require('backbone');
 var cdb = require('cartodb.js');
 var Router = require('../../../../../javascripts/cartodb/new_dashboard/router');
-var ImportsModel = require('../../../../../javascripts/cartodb/new_dashboard/background_importer/imports_model');
+var ImportsModel = require('../../../../../javascripts/cartodb/new_common/imports_model');
 var BackgroundImporterView = require('../../../../../javascripts/cartodb/new_dashboard/background_importer/background_importer_view');
 
 describe('new_dashboard/background_importer/background_importer_view', function() {
@@ -20,6 +20,7 @@ describe('new_dashboard/background_importer/background_importer_view', function(
       content_type: 'datasets'
     });
 
+    spyOn(cdb.god, 'bind').and.callThrough();
     spyOn(this.router.model, 'bind').and.callThrough();
 
     this.view = new BackgroundImporterView({
@@ -36,6 +37,18 @@ describe('new_dashboard/background_importer/background_importer_view', function(
     expect(this.view.$('.BackgroundImporter-header').length).toBe(1);
     expect(this.view.$('.BackgroundImporter-body').length).toBe(1);
     expect(this.view.$('.BackgroundImporter-list').length).toBe(1);
+  });
+
+  it('should listen for cdb.god events', function() {
+    var args = cdb.god.bind.calls.argsFor(0);
+    expect(args[0]).toEqual('datasetSelected remoteSelected');
+    expect(args[1]).toEqual(this.view._addDataset);
+    expect(args[2]).toEqual(this.view);
+
+    var args1 = cdb.god.bind.calls.argsFor(1);
+    expect(args1[0]).toEqual('fileDropped');
+    expect(args1[1]).toEqual(this.view._onDroppedFile);
+    expect(args1[2]).toEqual(this.view);
   });
 
   it('should toggle visibility depending the size of the collection', function() {
