+++ lib/assets/test/spec/cartodb/new_dashboard/background_importer/imports_model.spec.js
@@ -1,7 +1,7 @@
 var cdb = require('cartodb.js');
-var ImportsModel = require('../../../../../javascripts/cartodb/new_dashboard/background_importer/imports_model');
+var ImportsModel = require('../../../../../javascripts/cartodb/new_common/imports_model');
 var UploadModel = require('../../../../../javascripts/cartodb/new_common/upload_model');
-var ImportModel = require('../../../../../javascripts/cartodb/new_dashboard/background_importer/import_model');
+var ImportModel = require('../../../../../javascripts/cartodb/new_common/import_model');
 
 describe('new_dashboard/background_importer/imports_model', function() {
   beforeEach(function() {
